--- a/lib/cobra_commander.rb
+++ b/lib/cobra_commander.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
+require "cobra_commander/dependencies"
+require "cobra_commander/component"
+require "cobra_commander/umbrella"
+
 require "cobra_commander/cli"
 require "cobra_commander/cached_component_tree"
 require "cobra_commander/calculated_component_tree"
@@ -15,6 +19,13 @@ require "cobra_commander/executor"
 # testing components based on changes made.
 module CobraCommander
   UMBRELLA_APP_NAME = "App"
+
+  def self.umbrella(root_path, yarn: false, bundler: false, name: UMBRELLA_APP_NAME)
+    umbrella = Umbrella.new(name, root_path)
+    umbrella.add_source(:yarn, Dependencies::YarnWorkspace.new(root_path)) unless bundler
+    umbrella.add_source(:bundler, Dependencies::Bundler.new(root_path)) unless yarn
+    umbrella
+  end
 
   def self.umbrella_tree(path)
     CalculatedComponentTree.new(UMBRELLA_APP_NAME, path)