--- a/lib/cobra_commander/dependencies.rb
+++ b/lib/cobra_commander/dependencies.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "dependencies/yarn_workspace"
+require_relative "dependencies/bundler"