--- a/lib/cobra_commander/dependencies/yarn/package.rb
+++ b/lib/cobra_commander/dependencies/yarn/package.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+require "json"
+
+module CobraCommander
+  module Dependencies
+    module Yarn
+      class Package
+        attr_reader :path
+
+        def initialize(path)
+          @path = Pathname.new(File.join(path, "package.json")).realpath
+        end
+
+        def project_tag
+          name.match(%r{^@[\w-]+\/}).to_s
+        end
+
+        def name
+          json["name"]
+        end
+
+        def dependencies
+          json.fetch("dependencies", {})
+              .merge(json.fetch("devDependencies", {}))
+        end
+
+      private
+
+        def json
+          @json ||= JSON.parse(File.read(@path))
+        end
+      end
+    end
+  end
+end