--- a/lib/cobra_commander/umbrella.rb
+++ b/lib/cobra_commander/umbrella.rb
@@ -19,6 +19,7 @@ module CobraCommander
     end
 
     def resolve(component_root_path)
+      return root if root.root_paths.include?(component_root_path)
       components.find do |component|
         component.root_paths.include?(component_root_path)
       end