--- a/lib/cobra_commander/umbrella.rb
+++ b/lib/cobra_commander/umbrella.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module CobraCommander
+  # An umbrella application
   class Umbrella
     attr_reader :name, :path