--- a/lib/cobra_commander/version.rb
+++ b/lib/cobra_commander/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module CobraCommander
-  VERSION = "0.6.1"
+  VERSION = "0.7.0"
 end