--- a/spec/cobra_commander/cli_spec.rb
+++ b/spec/cobra_commander/cli_spec.rb
@@ -7,6 +7,24 @@ require "securerandom"
 RSpec.describe "cli", type: :aruba do
   before(:all) { @root = AppHelper.root }
 
+  describe "cobra do" do
+    it "executes the given command on all components" do
+      run_command_and_stop("cobra do --app #{@root} 'basename $PWD'", fail_on_error: true)
+
+      expect(last_command_started.output.split("\n").grep(/^[^=]/)).to match_array %w[
+        e
+        f
+        g
+        b
+        node_manifest
+        h
+        a
+        c
+        d
+      ]
+    end
+  end
+
   describe "checking the version" do
     it "reports the current version" do
       run_command_and_stop("cobra version", fail_on_error: true)