… .travis.yml
… CHANGELOG.md
… spec/cobra_commander/affected_spec.rb
… spec/cobra_commander/change_spec.rb
… spec/cobra_commander/executor_spec.rb
… spec/cobra_commander/graph_spec.rb
… spec/cobra_commander/umbrella_spec.rb