## Before
```bash
bundle exec cobra ls --format=list .  766.76s user 41.04s system 99% cpu 13:34.10 total
```

### After
```bash
bundle exec cobra ls --format=list .  2.07s user 1.34s system 90% cpu 3.775 total
```