@xjunior 2020-07-03
@benlangfeld this is in good shape now, 100% compatible with the previous implementation. I'll now update nitro to use this new version in its internal scripts and clean-up the cache handling stuff. I'll then create another PR with my suggestions for a new `cobra` CLI API.