@xjunior 2020-07-03
@benlangfeld sure thing, but before that I'll create a nitro branch with this version and see it getting tested/deployed to make sure it's all good.