@benlangfeld 2020-07-03
This appears to remove support for the cache, is that right? We should remove it from the interface and docs in that case, and make note of the backward incompatibility in the changelog.