@xjunior 2020-07-08
@benlangfeld this does not remove the cache options or CLI, but I'm adding a deprecation now. In my next PR I'll suggest a more complete API and remove all cache options.